'Jooperative' is a Q3A mod I created cirka 2006
The main parts of the mod, if I remember correctly, consisted of:
 * implementing cooperative gameplay (or something to that effect);
 * translating to Swedish;
 * replacing the boxing glove with a mixer;
 * replacing the rail gun with a sniper rifle.

I poked around with it and made some minor changes here and there, but
nothing major as far as I can tell. Unless you count useless,
inconsistent whitespace changes as major.

+++ code/renderer/tr_image.c
@@ -882,7 +882,7 @@ static void LoadBMP( const char *name, byte **pic, int *width, int *height )
 	{
 		ri.Error( ERR_DROP, "LoadBMP: only Windows-style BMP files supported (%s)\n", name );
 	}
-	if ( bmpHeader.fileSize != length )
+	if ( (int)bmpHeader.fileSize != length )
 	{
 		ri.Error( ERR_DROP, "LoadBMP: header size does not match file size (%d vs. %d) (%s)\n", bmpHeader.fileSize, length, name );
 	}
